Feature: Add function to reset widget (#15)
## Changelog
Add a function that allows to reset the current widget. This is useful
if a form is submitted, but the server returns an error, so the user is
forced to adjust their information. As this often doesn't reset the
page, the widget isn't automatically resetted. When submitting the form
again, this would cause an error as the captcha was 'already solved'.